requirements.txt
androguard==3.3.5
asn1crypto==1.5.1
asttokens==2.4.0
backcall==0.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==38.0.1
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.0
fonttools==4.43.1
future==0.18.3
idna==3.3
ipython==8.16.1
jedi==0.19.1
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
networkx==3.2
numpy==1.23.1
packaging==23.2
panda==0.3.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.39
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.19.0
pydot==1.4.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
six==1.16.0
stack-data==0.6.3
traitlets==5.11.2
urllib3==1.26.11
vboxapi==1.0
wcwidth==0.2.8
